requirements.txt
flask
flask-openapi3[swagger]
flask-limiter
requests
python-dotenv